Remove special casing of "first run"
I removed the special casing of the "first run" that used Exuberant Ctags
to directly create non-existing tag files. This additional code path
complicated the plug-in for an exceptional circumstance that wouldn't
be slow even if it was handled using the generic code. In other words,
out it goes in the quest for asynchronous tag updates!

                                * * *

These are the first steps towards implementing asynchronous tag updates
using only Vim as outlined in pull request #43:

  https://github.com/xolox/vim-easytags/pull/43